Moves instance group manager related classes to the main package.
