+++ src/main/java/gyro/google/compute/ComputeDistributionPolicyZoneConfiguration.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package gyro.google.compute.base;
+package gyro.google.compute;
 
 import com.google.api.services.compute.model.DistributionPolicyZoneConfiguration;
 import gyro.core.resource.Diffable;
