+++ src/main/java/gyro/google/compute/ComputeInstanceGroupManagerUpdatePolicy.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package gyro.google.compute.base;
+package gyro.google.compute;
 
 import java.util.Optional;
 
@@ -54,7 +54,7 @@ public class ComputeInstanceGroupManagerUpdatePolicy extends Diffable
     /**
      * The maximum number of instances that can be created above the specified targetSize during the update process. By default, a fixed value of 1 is used. This value can be either a fixed number or a percentage if the instance group has 10 or more instances. If you set a percentage, the number of instances will be rounded up if necessary.At least one of either maxSurge or maxUnavailable must be greater than 0. Learn more about maxSurge.
      *
-     * @subresource gyro.google.compute.base.ComputeFixedOrPercent
+     * @subresource gyro.google.compute.ComputeFixedOrPercent
      */
     public ComputeFixedOrPercent getMaxSurge() {
         return maxSurge;
@@ -67,7 +67,7 @@ public class ComputeInstanceGroupManagerUpdatePolicy extends Diffable
     /**
      * The maximum number of instances that can be unavailable during the update process. An instance is considered available if all of the following conditions are satisfied: - The instance's status is RUNNING. - If there is a health check on the instance group, the instance's liveness health check result must be HEALTHY at least once. If there is no health check on the group, then the instance only needs to have a status of RUNNING to be considered available.  By default, a fixed value of 1 is used. This value can be either a fixed number or a percentage if the instance group has 10 or more instances. If you set a percentage, the number of instances will be rounded up if necessary.At least one of either maxSurge or maxUnavailable must be greater than 0. Learn more about maxUnavailable.
      *
-     * @subresource gyro.google.compute.base.ComputeFixedOrPercent
+     * @subresource gyro.google.compute.ComputeFixedOrPercent
      */
     public ComputeFixedOrPercent getMaxUnavailable() {
         return maxUnavailable;
